omni-code reorganization and cleaning
